Fix table edition

+++ src/app/main/main.controller.coffee
@@ -2,10 +2,10 @@ angular.module 'iventureFront'
   .filter 'names', ->
     (input) ->
       types = []
-      types['Course'] = 'Cursos';
-      types['Track'] = 'Tracks';
-      types['Guest'] = 'Invitados';
-      types['Carousel'] = 'Aliados';
+      types['Course'] = 'Cursos'
+      types['Track'] = 'Tracks'
+      types['Guest'] = 'Invitados'
+      types['Carousel'] = 'Aliados'
 
       types[input]
 
@@ -18,7 +18,7 @@ angular.module 'iventureFront'
     {
       link: fn_link
     }
-  .controller 'AdminController', ($state) ->
+  .controller 'AdminController', ($state, LoopBackAuth) ->
     'ngInject'
     vm = this
     
@@ -28,8 +28,15 @@ angular.module 'iventureFront'
       {id: 'Guest', name: 'Invitados'},
       {id: 'Carousel', name: 'Aliados'}
     ]
+
+    logout = (event) ->
+      event.preventDefault()
+      LoopBackAuth.clearUser()
+      LoopBackAuth.clearStorage()
+      $state.go 'login'
+
     return
-  .controller 'CategoryController', ($state, toastr, $stateParams, Course, Track, Guest, Carousel) ->
+  .controller 'CategoryController', ($state, toastr, $stateParams, Course, Track, Guest, Carousel, $uibModal) ->
     'ngInject'
     vm = this
 
@@ -46,16 +53,37 @@ angular.module 'iventureFront'
 
     vm.delete = (event, id) ->
       event.preventDefault()
-      Get.deleteById({id: id})
-        .$promise
-        .then ->
-          toastr.success 'Elemento eliminado'
-          Get.find()
+
+      modal = $uibModal.open({
+        animation: true
+        template: """
+<div>
+  <div class="modal-header">
+    <h3 class="modal-title">Borrar elemento</h3>
+  </div>
+  <div class="modal-body">
+    <p>¿Deseas eliminar el elemento?</p>
+  </div>
+  <div class="modal-footer">
+    <button class="btn btn-danger" type="button" ng-click="$close(true)">Eliminar</button>
+    <button class="btn btn-default" type="button" ng-click="$close(false)">Cancelar</button>
+  </div>
+</div>
+"""
+        size: 'md'
+      })
+      modal.result.then (result) ->
+        if result
+          Get.deleteById({id: id})
             .$promise
-            .then (response) ->
-              vm.resources = response
-        .catch ->
-          toastr.error 'Error', 'No se pudo borrar el elemento'
+            .then ->
+              toastr.success 'Elemento eliminado'
+              Get.find()
+                .$promise
+                .then (response) ->
+                  vm.resources = response
+            .catch ->
+              toastr.error 'Error', 'No se pudo borrar el elemento'
   
     return
